Move getListValues from table_aws_ec2_instance.go to utils.go
